refactor the job directory to separate strategies from common operations
