+++ tests/test_scrape_metadata.py
@@ -5,7 +5,7 @@ import pandas as pd
 
 from db.mappings.article import Article
 from db.mappings.path import Path
-from job.steps.collaborative_filtering.scrape_metadata import scrape_upload_metadata
+from job.steps.scrape_metadata import scrape_upload_metadata
 from sites.helpers.requests import ArticleScrapingError, ScrapeFailure
 from sites.templates.site import Site
 from sites.sites import Sites
@@ -86,9 +86,9 @@ class TestScrapeMetadata(BaseTest):
         ]
         super().setUp()
 
-    @patch("job.steps.collaborative_filtering.warehouse.get_paths_to_update", return_value=NEW_VALID)
+    @patch("job.helpers.warehouse.get_paths_to_update", return_value=NEW_VALID)
     @patch(
-        "job.steps.collaborative_filtering.scrape_metadata.scrape_article",
+        "job.steps.scrape_metadata.scrape_article",
         return_value=VALID_SCRAPE,
     )
     def test_scrape_metadata__new_valid_recs(self, _, __) -> None:
@@ -96,9 +96,9 @@ class TestScrapeMetadata(BaseTest):
         res = Article.select().where(Article.site == self.site.name)
         assert len(res) == 3
 
-    @patch("job.steps.collaborative_filtering.warehouse.get_paths_to_update", return_value=EXISTING_VALID)
+    @patch("job.helpers.warehouse.get_paths_to_update", return_value=EXISTING_VALID)
     @patch(
-        "job.steps.collaborative_filtering.scrape_metadata.scrape_article",
+        "job.steps.scrape_metadata.scrape_article",
         return_value=VALID_SCRAPE,
     )
     def test_scrape_metadata__existing_valid_recs(self, _, __) -> None:
@@ -108,9 +108,9 @@ class TestScrapeMetadata(BaseTest):
         res = Article.select().where(Article.site == self.site.name)
         assert len(res) == 3
 
-    @patch("job.steps.collaborative_filtering.warehouse.get_paths_to_update", return_value=EXISTING_VALID)
+    @patch("job.helpers.warehouse.get_paths_to_update", return_value=EXISTING_VALID)
     @patch(
-        "job.steps.collaborative_filtering.scrape_metadata.scrape_article",
+        "job.steps.scrape_metadata.scrape_article",
         return_value=None,
         side_effect=scrape_error,
     )
@@ -123,9 +123,9 @@ class TestScrapeMetadata(BaseTest):
         res = Path.select().where(Path.site == self.site.name)
         assert len(res) == 0
 
-    @patch("job.steps.collaborative_filtering.warehouse.get_paths_to_update", return_value=EXISTING_VALID)
+    @patch("job.helpers.warehouse.get_paths_to_update", return_value=EXISTING_VALID)
     @patch(
-        "job.steps.collaborative_filtering.scrape_metadata.scrape_article",
+        "job.steps.scrape_metadata.scrape_article",
         return_value=None,
         side_effect=invalid_scrape_error,
     )
@@ -137,9 +137,9 @@ class TestScrapeMetadata(BaseTest):
         res = Article.select().where(Article.site == self.site.name)
         assert len(res) == 0
 
-    @patch("job.steps.collaborative_filtering.warehouse.get_paths_to_update", return_value=NEW_VALID)
+    @patch("job.helpers.warehouse.get_paths_to_update", return_value=NEW_VALID)
     @patch(
-        "job.steps.collaborative_filtering.scrape_metadata.scrape_article",
+        "job.steps.scrape_metadata.scrape_article",
         return_value=None,
         side_effect=safe_scrape_error,
     )
